+++ libs/blueprint/resourcehelpers/registry_test.go
@@ -7,8 +7,10 @@ import (
 
 	"github.com/newstack-cloud/bluelink/libs/blueprint/core"
 	"github.com/newstack-cloud/bluelink/libs/blueprint/errors"
+	"github.com/newstack-cloud/bluelink/libs/blueprint/internal/memstate"
 	"github.com/newstack-cloud/bluelink/libs/blueprint/provider"
 	"github.com/newstack-cloud/bluelink/libs/blueprint/schema"
+	"github.com/newstack-cloud/bluelink/libs/blueprint/state"
 	"github.com/newstack-cloud/bluelink/libs/blueprint/transform"
 	. "gopkg.in/check.v1"
 )
@@ -16,6 +18,7 @@ import (
 type RegistryTestSuite struct {
 	resourceRegistry Registry
 	testResource     *testExampleResource
+	stateContainer   state.Container
 }
 
 var _ = Suite(&RegistryTestSuite{})
@@ -42,11 +45,13 @@ func (s *RegistryTestSuite) SetUpTest(c *C) {
 		},
 	}
 
+	s.stateContainer = memstate.NewMemoryStateContainer()
 	s.testResource = testRes.(*testExampleResource)
 	s.resourceRegistry = NewRegistry(
 		providers,
 		transformers,
 		time.Millisecond,
+		s.stateContainer,
 		/* params */ nil,
 	)
 }
@@ -169,6 +174,182 @@ func (s *RegistryTestSuite) Test_deploy_resource(c *C) {
 			"spec.id": core.MappingNodeFromString("test-example-resource-item-id-1"),
 		},
 	})
+}
+
+func (s *RegistryTestSuite) Test_look_up_resource_in_state_by_external_id(c *C) {
+	persistedResourceState := &state.ResourceState{
+		ResourceID: "test-resource-id-1",
+		Name:       "testResource1",
+		Type:       "test/exampleResource",
+		SpecData: &core.MappingNode{
+			Fields: map[string]*core.MappingNode{
+				"id": core.MappingNodeFromString("test-example-resource-item-id-1"),
+			},
+		},
+		Metadata: &state.ResourceMetadataState{},
+	}
+	err := s.stateContainer.Instances().Save(
+		context.TODO(),
+		state.InstanceState{
+			InstanceID:   "test-blueprint-id",
+			InstanceName: "TestBlueprint",
+			Status:       core.InstanceStatusDeployed,
+			ResourceIDs: map[string]string{
+				"testResource1": "test-resource-id-1",
+			},
+			Resources: map[string]*state.ResourceState{
+				"testResource1": persistedResourceState,
+			},
+		},
+	)
+	c.Assert(err, IsNil)
+
+	resource, err := s.resourceRegistry.LookupResourceInState(
+		context.TODO(),
+		&provider.ResourceLookupInput{
+			InstanceID:   "test-blueprint-id",
+			ResourceType: "test/exampleResource",
+			ExternalID:   "test-example-resource-item-id-1",
+		},
+	)
+	c.Assert(err, IsNil)
+
+	c.Assert(
+		resource.ResourceID,
+		Equals,
+		persistedResourceState.ResourceID,
+	)
+}
+
+func (s *RegistryTestSuite) Test_look_up_for_missing_resource_in_state_by_external_id_returns_nil(c *C) {
+	persistedResourceState := &state.ResourceState{
+		ResourceID: "test-resource-id-1",
+		Name:       "testResource1",
+		Type:       "test/exampleResource",
+		SpecData: &core.MappingNode{
+			Fields: map[string]*core.MappingNode{
+				"id": core.MappingNodeFromString("test-example-resource-item-id-1"),
+			},
+		},
+		Metadata: &state.ResourceMetadataState{},
+	}
+	err := s.stateContainer.Instances().Save(
+		context.TODO(),
+		state.InstanceState{
+			InstanceID:   "test-blueprint-id",
+			InstanceName: "TestBlueprint",
+			Status:       core.InstanceStatusDeployed,
+			ResourceIDs: map[string]string{
+				"testResource1": "test-resource-id-1",
+			},
+			Resources: map[string]*state.ResourceState{
+				"testResource1": persistedResourceState,
+			},
+		},
+	)
+	c.Assert(err, IsNil)
+
+	resource, err := s.resourceRegistry.LookupResourceInState(
+		context.TODO(),
+		&provider.ResourceLookupInput{
+			InstanceID:   "test-blueprint-id",
+			ResourceType: "test/exampleResource",
+			ExternalID:   "test-example-resource-item-id-missing",
+		},
+	)
+	c.Assert(err, IsNil)
+
+	c.Assert(resource, IsNil)
+}
+
+func (s *RegistryTestSuite) Test_check_resource_in_state_by_external_id(c *C) {
+	persistedResourceState := &state.ResourceState{
+		ResourceID: "test-resource-id-1",
+		Name:       "testResource1",
+		Type:       "test/exampleResource",
+		SpecData: &core.MappingNode{
+			Fields: map[string]*core.MappingNode{
+				"id": core.MappingNodeFromString("test-example-resource-item-id-1"),
+			},
+		},
+		Metadata: &state.ResourceMetadataState{},
+	}
+	err := s.stateContainer.Instances().Save(
+		context.TODO(),
+		state.InstanceState{
+			InstanceID:   "test-blueprint-id",
+			InstanceName: "TestBlueprint",
+			Status:       core.InstanceStatusDeployed,
+			ResourceIDs: map[string]string{
+				"testResource1": "test-resource-id-1",
+			},
+			Resources: map[string]*state.ResourceState{
+				"testResource1": persistedResourceState,
+			},
+		},
+	)
+	c.Assert(err, IsNil)
+
+	hasResource, err := s.resourceRegistry.HasResourceInState(
+		context.TODO(),
+		&provider.ResourceLookupInput{
+			InstanceID:   "test-blueprint-id",
+			ResourceType: "test/exampleResource",
+			ExternalID:   "test-example-resource-item-id-1",
+		},
+	)
+	c.Assert(err, IsNil)
+
+	c.Assert(
+		hasResource,
+		Equals,
+		true,
+	)
+}
+
+func (s *RegistryTestSuite) Test_check_missing_resource_in_state_by_external_id_returns_false(c *C) {
+	persistedResourceState := &state.ResourceState{
+		ResourceID: "test-resource-id-1",
+		Name:       "testResource1",
+		Type:       "test/exampleResource",
+		SpecData: &core.MappingNode{
+			Fields: map[string]*core.MappingNode{
+				"id": core.MappingNodeFromString("test-example-resource-item-id-1"),
+			},
+		},
+		Metadata: &state.ResourceMetadataState{},
+	}
+	err := s.stateContainer.Instances().Save(
+		context.TODO(),
+		state.InstanceState{
+			InstanceID:   "test-blueprint-id",
+			InstanceName: "TestBlueprint",
+			Status:       core.InstanceStatusDeployed,
+			ResourceIDs: map[string]string{
+				"testResource1": "test-resource-id-1",
+			},
+			Resources: map[string]*state.ResourceState{
+				"testResource1": persistedResourceState,
+			},
+		},
+	)
+	c.Assert(err, IsNil)
+
+	hasResource, err := s.resourceRegistry.HasResourceInState(
+		context.TODO(),
+		&provider.ResourceLookupInput{
+			InstanceID:   "test-blueprint-id",
+			ResourceType: "test/exampleResource",
+			ExternalID:   "test-example-resource-item-id-missing",
+		},
+	)
+	c.Assert(err, IsNil)
+
+	c.Assert(
+		hasResource,
+		Equals,
+		false,
+	)
 }
 
 func (s *RegistryTestSuite) Test_produces_error_for_missing_provider(c *C) {